less than core functionality
untested
Should support:
1) extracting parts to STDOUT by offset&length.
2) parsing the header and all XCRFAT (word creation of mine) entries
3) comparing external files to parts inside by offset

API should support:
4) appending new entries where
a) the length is known in advance.
b) the length will be known after write operation finished.
c) it is necessary to extend the XCRFAT and conditionally move the data of the existing entry having the lowest offset to the end of the file (also moves more entries if necessary).